- Get rid of __LIBRETRO__ ifdef - Turn cur_dir_slen into size_t

+++ libretro/winui.h
@@ -7,7 +7,7 @@
 extern	uint8_t	Debug_Text, Debug_Grp, Debug_Sp;
 
 extern char cur_dir_str[];
-extern int cur_dir_slen;
+extern size_t cur_dir_slen;
 extern int speedup_joy[0xff];
 
 void WinUI_Init(void);
